% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{collect.odsql}
\title{Generics. Collect ODSQL object and build the API query.}
\usage{
collect(.data, ...)

\method{collect}{odsql}(.data, ...)
}
\description{
Generics. Collect ODSQL object and build the API query.

Collect ODSQL object and build the API query
}
