% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain.R
\name{domain}
\alias{domain}
\alias{set_domain}
\alias{unset_domain}
\title{API domain as global option}
\usage{
set_domain(domain)

unset_domain()

domain()
}
\arguments{
\item{domain}{An API domain passed as a string.}
}
\value{
\itemize{
\item \code{set_domain()}: a named list with key \code{odsql_api_domain}, and value the
domain set value.
\item \code{unset_domain()}: a named list with \code{odsql_api_domain} and domain value
just before it is unset.
\item \code{domain()}: a string with the set domain value, or \code{NULL} if not set.
}
}
\description{
These functions allow to manage the global option \code{odsql_api_domain}.
\itemize{
\item \code{set_domain()}: set a domain as a global option.
\item \code{unset_domain()}: unset a domain.
\item \code{domain()}: get the domain option.
}
}
