% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odsql.R
\name{odsql}
\alias{odsql}
\title{Creates an Opendatasoft Query Language (odsql) object}
\usage{
odsql(dataset_id)
}
\arguments{
\item{dataset_id}{Character, identifier of the dataset}
}
\description{
Creates an Opendatasoft Query Language (odsql) object
}
