% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\alias{rename.odsql}
\title{Generics. Rename columns in an odsql query.}
\usage{
rename(.data, ...)

\method{rename}{odsql}(.data, ...)
}
\description{
Generics. Rename columns in an odsql query.
}
