% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.R
\name{req}
\alias{req}
\title{Create a new HTTP request}
\usage{
req(.res, ..., .dom = domain(), .body = NULL, .headers = req_headers())
}
\arguments{
\item{.res}{A resource (res) URL as a string. This string supports embedding
of R variable names in curly braces whose values are looked up in parameter
names supplied in \code{...} and interpolated.}

\item{...}{Name value pairs specifying query components or parameters.}

\item{.dom}{Opendatasoft's Explore API domain (dom). The Explore API is
accessed using a base URL that is specific to a domain. Defaults to
global option returned by \code{\link[=domain]{domain()}}.}

\item{.body}{A literal string or raw vector to send as body.}

\item{.headers}{An S3 list with class \code{odsql_req_hdr}. Use the helper
\code{\link[=req_headers]{req_headers()}} to create such an object.}
}
\value{
An HTTP request: an S3 list with class \code{httr2_request}.
}
\description{
\code{\link[=req]{req()}} creates an HTTP request object.
}
\examples{
\dontrun{

# Explicitly pass the API domain (`.dom`) to `req()`.
req(
  .dom = "documentation-resources.opendatasoft.com",
  .res = "/catalog/datasets/{dataset_id}",
  dataset_id = "roman-emperors",
  limit = 10
  )

# Or define the API domain as a global option and omit `.dom` from subsequent
# calls to `req()`.
set_domain("documentation-resources.opendatasoft.com")
req(.res = "/catalog/datasets/{dataset_id}", dataset_id = "roman-emperors", limit = 10)
}

}
\keyword{internal}
