% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\alias{select.odsql}
\title{Generics. Select columns in an odsql query.}
\usage{
select(.data, ...)

\method{select}{odsql}(.data, ...)
}
\description{
Generics. Select columns in an odsql query.
}
